+++ Classes/00-Base/NuGetPackageDefinition.ps1
@@ -1,0 +1,13 @@
+class NuGetPackageDefinition : Package {
+    [string]$FeedUrl
+    [string]$DownloadUri
+    [string]$InstalledVersion
+    [string]$Description
+    [string]$Authors
+
+    NuGetPackageDefinition() {
+        $this.Provider = 'NuGet'
+        $this.InstallerType = 'NuGet'
+        $this.RequiresElevation = $false
+    }
+}