+++ Public/Start-PhoenixRestoreJob.ps1
@@ -24,13 +24,15 @@ function Start-PhoenixRestoreJob {
             'WinGet',
             'Chocolatey',
             'Scoop',
-            'PowerShell Gallery'
+            'PowerShell Gallery',
+            'NuGet'
         )]
         [string[]]$Provider = @(
             'WinGet',
             'Chocolatey',
             'Scoop',
-            'PowerShell Gallery'
+            'PowerShell Gallery',
+            'NuGet'
         ),
 
         [Parameter()]